Huge Reddit model refactor
